requirements.txt
Django==4.0.1
mysqlclient
djangorestframework
django-rest-knox
django-pagination
djangorestframework_simplejwt
-- NOTE: install without sudo